finished swagger and change create in base
