+++ request.rest
@@ -1,32 +1,39 @@
+### JWT Token 
+@jwtToken=Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NzgzY2ZjMTZhNzVmZWFkZTQzZWVkMzEiLCJyYW5kb20iOiIwLjkxNDE0OTA0NjAxODIxODkiLCJpYXQiOjE3MzY2OTE2NTIsImV4cCI6MTczNjcwMjQ1Mn0.unb_Qwr_4Y5ff70GY5wDomrTXp_Rg3sRSatvQNtqWPE
+
 
 ###########################################################
 ## Posts route
 ###########################################################
 
 GET http://localhost:3000/posts
+Authorization: {{jwtToken}}
 
 ###
 
 GET http://localhost:3000/posts?sender=67815b9697a75440c50555cd
+Authorization: {{jwtToken}}
 
 ###
 
 GET http://localhost:3000/posts/673f93a854f16b4155f3c18d
+Authorization: {{jwtToken}}
 
 ###
 
 POST http://localhost:3000/posts
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
     "title": "sdsdirs",
-    "sender": "67815b9697a75440c50555cd",
     "content": "Shir is the best"
 }
 
 ###
 
 PUT http://localhost:3000/posts/677f9785b1032f6d87baf29e
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -37,28 +44,32 @@ Content-Type: application/json
 ### delete post and all its comments
 
 DELETE http://localhost:3000/posts/677fe9de62b3029664f9903c
+Authorization: {{jwtToken}}
 
 ###########################################################
 ## Comments route
 ###########################################################
 
 POST http://localhost:3000/comments
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
 
-    "postId": "677fe9de62b3029664f9903c",
-    "sender": "67815b9697a75440c50555cd",
+    "postId": "6782c77fd4ee7407d84065b2",
+    "sender": "mich",
     "content": "2"
 }
 
 ###
 
 GET http://localhost:3000/comments
+Authorization: {{jwtToken}}
 
 ###
 
 PUT http://localhost:3000/comments/677d7cdd4133bd4c06d6a27c
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -68,23 +79,28 @@ Content-Type: application/json
 ###
 
 GET http://localhost:3000/comments/677d76c31815bd58518421b8
+Authorization: {{jwtToken}}
 
 ###
 DELETE http://localhost:3000/comments/677f9785b1032f6d87baf29e
+Authorization: {{jwtToken}}
 
 ### get all comments related to post id
 
 GET http://localhost:3000/comments?postId=677fe9de62b3029664f9903c
+Authorization: {{jwtToken}}
 
 ###
 
 GET http://localhost:3000/comments?sender=67815b9697a75440c50555cd&postId=677d767e1123815bd58518421b6
+Authorization: {{jwtToken}}
 
 ###########################################################
 ## Auth route
 ###########################################################
 
 POST http://localhost:3000/auth/register 
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -110,19 +126,24 @@ Content-Type: application/json
 
 ### Get all users
 GET http://localhost:3000/users
+Authorization: {{jwtToken}}
 
 ### Get user by id
 GET http://localhost:3000/users/677ea922c6682ae9e754e256
+Authorization: {{jwtToken}}
 
 ### Get user by username
 GET http://localhost:3000/users?username=mich
+Authorization: {{jwtToken}}
 
 ### Get user by email
 GET http://localhost:3000/users?email=bla@c.com
+Authorization: {{jwtToken}}
 
 ### Create new user
 
 POST http://localhost:3000/users
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -134,6 +155,7 @@ Content-Type: application/json
 ### Update username
 
 PUT http://localhost:3000/users/677ea922c6682ae9e754e256
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -143,6 +165,7 @@ Content-Type: application/json
 ### Update email
 
 PUT http://localhost:3000/users/677ea922c6682ae9e754e256
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -152,6 +175,7 @@ Content-Type: application/json
 ### Update password
 
 PUT http://localhost:3000/users/677ea922c6682ae9e754e256
+Authorization: {{jwtToken}}
 Content-Type: application/json
 
 {
@@ -161,4 +185,4 @@ Content-Type: application/json
 ### Delete user
 
 DELETE http://localhost:3000/users/677fab696ce534642d6349bd
-
+Authorization: {{jwtToken}}
