+++ src/controllers/postsController.ts
@@ -11,35 +11,6 @@ class PostsController extends BaseController<IPost> {
     super(postModel);
   }
 
-  async create(req: Request, res: Response): Promise<void> {
-    try {
-      const sender = req.body.sender;
-
-      if (sender) {
-        if (mongoose.Types.ObjectId.isValid(sender)) {
-          const user = await userModel.findOne({
-            _id: new mongoose.Types.ObjectId(sender),
-          });
-          if (!user) {
-            res.status(status.NOT_FOUND).send('User not found');
-            return;
-          }
-        }
-      } else {
-        const userId = req.body.payload.userId;
-        const post = {
-          ...req.body,
-          sender: userId,
-        };
-        req.body = post;
-      }
-
-      super.create(req, res);
-    } catch (error) {
-      res.status(status.BAD_REQUEST).send(error);
-    }
-  }
-
   async deleteItem(req: Request, res: Response): Promise<void> {
     try {
       const postId = req.params.id;
