+++ src/routes/commentsRoute.ts
@@ -51,12 +51,21 @@ const router = Router();
  * /comments:
  *   get:
  *     summary: Get all comments
- *     description: Returns the list of all the comments
+ *     description: Returns the list of all the comments or filterd by postId
  *     tags:
  *       - Comments
+ *     security:
+ *       - bearerAuth: []
+ *     parameters:
+ *      - in: query
+ *        name: postId
+ *        schema:
+ *          type: string
+ *        required: false
+ *        description: The ID of the post to filter comments
  *     responses:
  *       '200':
- *         description: A list of the comments
+ *         description: A list of the comments or filterd
  *         content:
  *           application/json:
  *             schema:
@@ -76,6 +85,8 @@ router.get('/', commentsController.getAll.bind(commentsController));
  *     description: Returns single comment
  *     tags:
  *       - Comments
+ *     security:
+ *       - bearerAuth: []
  *     parameters:
  *      - in: path
  *        name: id
@@ -85,7 +96,7 @@ router.get('/', commentsController.getAll.bind(commentsController));
  *        description: ID of the comment
  *     responses:
  *       '200':
- *         description: A comments
+ *         description: A comment
  *         content:
  *           application/json:
  *             schema:
@@ -99,10 +110,118 @@ router.get('/', commentsController.getAll.bind(commentsController));
  */
 router.get('/:id', commentsController.getById.bind(commentsController));
 
+/**
+ * @swagger
+ * /comments:
+ *   post:
+ *     summary: Create a new comment
+ *     description: Create a new comment
+ *     tags:
+ *       - Comments
+ *     security:
+ *       - bearerAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               postId:
+ *                 type: string
+ *                 description: postId of the post to comment on
+ *               sender:
+ *                 type: string
+ *                 description: The sender of the comment
+ *               content:
+ *                 type: string
+ *                 description: The content of the comment
+ *             required:
+ *               - postId
+ *               - sender
+ *               - content
+ *     responses:
+ *       '201':
+ *         description: Comment created successfully
+ *         content:
+ *           application/json:
+ *             schema:
+ *                 $ref: '#/components/schemas/Comment'
+ *       '404':
+ *         description: Post not found
+ *       '500':
+ *         description: Server error
+ */
 router.post('/', commentsController.create.bind(commentsController));
 
+/**
+ * @swagger
+ * /comments/{id}:
+ *   put:
+ *     summary: Update a comment
+ *     description: Update a comment
+ *     tags:
+ *       - Comments
+ *     security:
+ *       - bearerAuth: []
+ *     parameters:
+ *      - in: path
+ *        name: id
+ *        schema:
+ *          type: string
+ *        required: true
+ *        description: ID of the comment
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               content:
+ *                 type: string
+ *                 description: The new content of the comment
+ *             required:
+ *               - content
+ *     responses:
+ *       '200':
+ *         description: Comment updated successfully
+ *         content:
+ *           application/json:
+ *             schema:
+ *                 $ref: '#/components/schemas/Comment'
+ *       '404':
+ *         description: Comment not found
+ *       '500':
+ *         description: Server error
+ */
 router.put('/:id', commentsController.updateItem.bind(commentsController));
 
+/**
+ * @swagger
+ * /comments/{id}:
+ *   delete:
+ *     summary: Delete a comment by ID
+ *     description: Delete a single comment by its ID
+ *     tags:
+ *       - Comments
+ *     security:
+ *       - bearerAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         schema:
+ *           type: string
+ *         required: true
+ *         description: The ID of the comment
+ *     responses:
+ *       '200':
+ *         description: Comment deleted successfully
+ *       '404':
+ *         description: Comment not found
+ *       '500':
+ *         description: Server error
+ */
 router.delete('/:id', commentsController.deleteItem.bind(commentsController));
 
 export default router;
